backend/requirements.txt
django
djangorestframework
mimesis
channels
daphne